move Expressions namespace to actual files
